Fix remaining warnings: cast-qual, sign-conversion, unused-parameter, pedantic
- Suppress unavoidable VIO_STR/string-literal cast-qual warnings with
  pragmas in volumes.c, output_mnc.c, files.c, test files
- Fix sign-conversion in m2util.c, value_conversion.c, minc_compat.c
- Fix unused-parameter in nd_loop.c and value_conversion.c
- Fix -Wpedantic empty translation unit in minc2_error.c
- Suppress -Wformat-nonliteral in minc_error.c (format string is a parameter)
- Fix int/size_t conversions in ezminc templates

Reduces gcc-15 warnings from 2005 to 43 (remaining are inherent to
VIO_STR being char* while MINC macros expand to string literals).

+++ ezminc/minc_1_rw.cpp
@@ -728,7 +728,7 @@ namespace minc
         case dim_info::DIM_VEC:_info[ui].name=MIvector_dimension;_map_to_std[0]=i;break;
         //default: REPORT_ERROR("Unknown Dimension!");
       }
-      CHECK_MINC_CALL(mdims[i]=ncdimdef(_mincid, _info[ui].name.c_str(), _info[ui].length));
+      CHECK_MINC_CALL(mdims[i]=ncdimdef(_mincid, _info[ui].name.c_str(), static_cast<long>(_info[ui].length)));
       if(_info[ui].dim!=dim_info::DIM_VEC)
       {
         int dimid;
